enable CI for Mac. build, launch, and test automatically on macos.

+++ run_all_tests.sh
@@ -37,8 +37,14 @@ else
   XDISPLAY=""
 fi
 
-tools/formatters/enforce_clang_format.sh check_only
-tools/formatters/enforce_qml_format.sh check_only
+if [[ "$OSTYPE" != "darwin"* ]]; then
+  # TODO: we can make it work on mac with clang-format-10
+  tools/formatters/enforce_clang_format.sh check_only
+
+  # TODO: we already made this work in a different repo. learn from it.
+  tools/formatters/enforce_qml_format.sh check_only
+fi
+
 
 ./build_app.sh
 
@@ -46,10 +52,12 @@ tools/formatters/enforce_qml_format.sh check_only
 tools/auto_test/run_cpp_auto_tests.sh
 
 if [[ -n ${XDISPLAY-} ]]; then
-  # 'coproc' so that SIGINT will still work (https://unix.stackexchange.com/a/478697/11592)
-  coproc Xvfb ${XDISPLAY} -screen 0 1024x768x16
-  VIRT_FB_PID=$!
-  sleep 4 # time to (probabilistically) ensure that Xvfb has started
+  if [[ "$OSTYPE" != "darwin"* ]]; then
+    # 'coproc' so that SIGINT will still work (https://unix.stackexchange.com/a/478697/11592)
+    coproc Xvfb ${XDISPLAY} -screen 0 1024x768x16
+    VIRT_FB_PID=$!
+    sleep 4 # time to (probabilistically) ensure that Xvfb has started
+  fi
 fi
 
 # run gui tests which execute the actual app binary:
@@ -61,5 +69,7 @@ if [[ -z ${MYAPP_TEMPLATE_SKIP_APPIMAGE-} ]]; then
 fi
 
 if [[ -n ${XDISPLAY-} ]]; then
-   kill -SIGINT $VIRT_FB_PID
+  if [[ "$OSTYPE" != "darwin"* ]]; then
+    kill -SIGINT $VIRT_FB_PID
+  fi
 fi
